+++ components/file-explorer/viewers/base.tsx
@@ -84,10 +84,10 @@ export function BaseViewer({
                 return <LargeViewSelect setType={setType} />
             default: {
                 const Viewer = VIEWERS.find((v) => v.type === tab.viewerType)!.component
-                return <Viewer data={data} tab={tab} updateTab={updateTab} />
+                return <Viewer data={error ? undefined : data} tab={tab} updateTab={updateTab} />
             }
         }
-    }, [data, setType, tab, updateTab])
+    }, [data, error, setType, tab, updateTab])
 
     return (
         <>
@@ -97,7 +97,7 @@ export function BaseViewer({
                     <LoaderCircle className="size-4 animate-spin text-muted-foreground" />
                 </div>
             ) : (
-                <div className="grow">{Content}</div>
+                <div className="grow max-h-full max-w-full min-h-0 overflow-auto">{Content}</div>
             )}
             <div className="flex justify-between items-center bg-muted/50 text-sm text-muted-foreground px-2">
                 <div className="truncate">Path: {tab.filePath}</div>