+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.13.0] - 2026-07-13
+
+### Added
+- Added support for the RO-Crate v1.3.0 specification
+- Added help tooltips to the AI Assistant configuration modal
+- The graph view now remembers the last viewport position and node layout for the duration of the session
+- Added a highlight to all edges connected to the selected entity in the graph view
+- Added a simple CSV viewer to the file explorer
+- Added options for exporting a crate without compression (especially useful for large crates)
+
+### Changed
+- The text viewer is now shown for any file with type text/* (if there is no other default)
+- The main menu now allows creating new RO-Crates from (multiple) files
+- Updated the documentation modal to add links to guides on GitHub and the YouTube tutorial series
+
+### Fixed
+- Creating an RO-Crate from a file or folder with an existing ro-crate-metadata.json now works correctly
+
 ## [1.12.0] - 2026-06-25
 
 ### Added